+++ infrastructure/entry-points/reactive-web/src/main/java/co/com/pragma/api/dto/RoleResponseDto.java
@@ -1,4 +1,4 @@
 package co.com.pragma.api.dto;
 
-public record RoleResponseDto(Long idRol, String nombre, String descripcion) {
+public record RoleResponseDto(Integer idRol, String nombre, String descripcion) {
 }
